Add new tests regarding generate features. Direct stdout and stderr to separate files to avoid corruption. Refactor DevTest to move utility methods to a base class.
